world coordinates float => double
Turn all world coordinates from float into double.
Float precision is not good enough with 32-bit keys when operating
nearer the edge of the map.
Double precision is actually faster on most modern processors.

Fix broken test cases by this change, due to differences in float
rounding giving different OcTreeKey's. Move such coordinates away
from the resolution boundary. Relax some tree size constraints for
radially constructed trees to allow for small variations in compilers
and architecture. Use FLOAT_EQ where appropriate in test cases.

Also, fix the ray test to not project a ray from the center of the
world to the edge, which on a 32-bit map is a long way off. Now that
there is double precision, adjust the coordinates from the edge of
the world slightly towards the center and test casting the ray into
the boundary from near the boundary, resulting in a test with
reasonable runtime.

Finally, fix the test macros to keep order of operations by adding
parenthesis around the macro arguments.

+++ octomap/include/octomap/ColorOcTree.h
@@ -129,8 +129,8 @@ namespace octomap {
     ColorOcTreeNode* setNodeColor(const OcTreeKey& key, uint8_t r, 
                                  uint8_t g, uint8_t b);
 
-    ColorOcTreeNode* setNodeColor(float x, float y, 
-                                 float z, uint8_t r, 
+    ColorOcTreeNode* setNodeColor(double x, double y, 
+                                 double z, uint8_t r, 
                                  uint8_t g, uint8_t b) {
       OcTreeKey key;
       if (!this->coordToKeyChecked(point3d(x,y,z), key)) return NULL;
@@ -141,8 +141,8 @@ namespace octomap {
     ColorOcTreeNode* averageNodeColor(const OcTreeKey& key, uint8_t r, 
                                   uint8_t g, uint8_t b);
     
-    ColorOcTreeNode* averageNodeColor(float x, float y, 
-                                      float z, uint8_t r, 
+    ColorOcTreeNode* averageNodeColor(double x, double y, 
+                                      double z, uint8_t r, 
                                       uint8_t g, uint8_t b) {
       OcTreeKey key;
       if (!this->coordToKeyChecked(point3d(x,y,z), key)) return NULL;
@@ -153,8 +153,8 @@ namespace octomap {
     ColorOcTreeNode* integrateNodeColor(const OcTreeKey& key, uint8_t r, 
                                   uint8_t g, uint8_t b);
     
-    ColorOcTreeNode* integrateNodeColor(float x, float y, 
-                                      float z, uint8_t r, 
+    ColorOcTreeNode* integrateNodeColor(double x, double y, 
+                                      double z, uint8_t r, 
                                       uint8_t g, uint8_t b) {
       OcTreeKey key;
       if (!this->coordToKeyChecked(point3d(x,y,z), key)) return NULL;
