+++ octomap/src/math/Quaternion.cpp
@@ -52,7 +52,7 @@ namespace octomath {
     data[3] = other(3);
   }
 
-  Quaternion::Quaternion(float uu, float xx, float yy, float zz) {
+  Quaternion::Quaternion(double uu, double xx, double yy, double zz) {
     u() = uu;
     x() = xx;
     y() = yy;
@@ -77,10 +77,10 @@ namespace octomath {
       {    -spitch,                  cpitch*sroll,                  cpitch*croll}
     };
 
-    float _u = (float) (sqrt(std::max(0., 1 + m[0][0] + m[1][1] + m[2][2]))/2.0);
-    float _x = (float) (sqrt(std::max(0., 1 + m[0][0] - m[1][1] - m[2][2]))/2.0);
-    float _y = (float) (sqrt(std::max(0., 1 - m[0][0] + m[1][1] - m[2][2]))/2.0);
-    float _z = (float) (sqrt(std::max(0., 1 - m[0][0] - m[1][1] + m[2][2]))/2.0);
+    double _u = (sqrt(std::max(0., 1 + m[0][0] + m[1][1] + m[2][2]))/2.0);
+    double _x = (sqrt(std::max(0., 1 + m[0][0] - m[1][1] - m[2][2]))/2.0);
+    double _y = (sqrt(std::max(0., 1 - m[0][0] + m[1][1] - m[2][2]))/2.0);
+    double _z = (sqrt(std::max(0., 1 - m[0][0] - m[1][1] + m[2][2]))/2.0);
     u() = _u;
     x() = (m[2][1] - m[1][2])>=0?fabs(_x):-fabs(_x);
     y() = (m[0][2] - m[2][0])>=0?fabs(_y):-fabs(_y);
@@ -90,21 +90,21 @@ namespace octomath {
   Quaternion::Quaternion(const Vector3& axis, double angle) {
     double sa = sin(angle/2);
     double ca = cos(angle/2);
-    x() = (float) (axis.x()*sa);
-    y() = (float) (axis.y()*sa);
-    z() = (float) (axis.z()*sa);
-    u() = (float) ca;
+    x() = (axis.x()*sa);
+    y() = (axis.y()*sa);
+    z() = (axis.z()*sa);
+    u() = ca;
   }
 
-  float Quaternion::norm () const {
+  double Quaternion::norm () const {
     double n = 0;
     for (unsigned int i=0; i<4; i++) {
       n += operator()(i) * operator()(i);
     }
-    return (float) sqrt(n);
+    return sqrt(n);
   }
 
-  void Quaternion::operator/= (float x) 
+  void Quaternion::operator/= (double x) 
   { 
     for (unsigned int i=0; i<4; ++i)
       operator()(i) /= x;
@@ -156,9 +156,9 @@ namespace octomath {
     m[2][1] = yz + ux;
     m[1][2] = yz - ux;
 
-    float roll  = (float) atan2(m[2][1], m[2][2]);
-    float pitch = (float) atan2(-m[2][0], sqrt(m[2][1]*m[2][1] + m[2][2]*m[2][2]));
-    float yaw   = (float) atan2(m[1][0], m[0][0]);
+    double roll  = atan2(m[2][1], m[2][2]);
+    double pitch = atan2(-m[2][0], sqrt(m[2][1]*m[2][1] + m[2][2]*m[2][2]));
+    double yaw   = atan2(m[1][0], m[0][0]);
 
     return Vector3(roll, pitch, yaw);
   }
@@ -234,7 +234,7 @@ namespace octomath {
   Quaternion& Quaternion::normalize (){
     double len = norm ();
     if (len > 0)
-      *this /= (float) len;
+      *this /= len;
     return *this;
   }
 
@@ -282,7 +282,7 @@ namespace octomath {
     double val = 0;
     for (unsigned int i=0; i<4; i++) {
       s.read((char*)&val, sizeof(val));
-      operator()(i) = (float) val;
+      operator()(i) = val;
     }
     return s;
   }
